Add pairify_map
By default `g:pairify_map` is set to "<C-J>", consistent with the older
mapping, however if it is set to a single character value, in case when
no match is found, we simply echo the value itself to type it literally.

This is in relation to #1, so allowing to set it to a value such as `;`
so typing `;` by itself when no matching is applicable would type the
character itself.